docs: 📝 reorganize agent guides and sync translations
