+++ docker-compose.security.yml
@@ -29,6 +29,8 @@ services:
         environment:
             - DOCKER_CONTAINER=true
             - NODE_ENV=production
+            - MCP_TRANSPORT=http
+            - MCP_BEARER=
             - ALLOWED_ORIGINS=https://example.com,https://*.example.org
             - MAX_SCREENSHOT_SIZE=2097152
             - MAX_CONTENT_LENGTH=1048576
@@ -48,6 +50,9 @@ services:
         # Red aislada (sin acceso externo directo)
         networks:
             - isolated
+
+        ports:
+            - "3333:3333"
 
         # Healthcheck
         healthcheck:
