+++ package.json
@@ -13,10 +13,12 @@
 		"README.md"
 	],
 	"scripts": {
-		"build": "tsc && chmod +x dist/*.js",
-		"dev": "tsc --watch",
-		"start": "node dist/index.js",
-		"test": "jest",
+                "build": "tsc && chmod +x dist/index.js || true",
+                "dev": "tsc --watch",
+                "start": "node dist/index.js",
+                "start:http": "MCP_TRANSPORT=http node dist/index.js",
+                "dev:http": "MCP_TRANSPORT=http PORT=3333 tsx src/index.ts",
+                "test": "jest",
 		"test:watch": "jest --watch",
 		"test:coverage": "jest --coverage",
 		"lint": "eslint src --ext .ts --fix",
@@ -28,13 +30,16 @@
 		"release": "semantic-release"
 	},
 	"dependencies": {
-		"@modelcontextprotocol/sdk": "^1.0.1",
-		"puppeteer": "^23.4.0"
-	},
-	"devDependencies": {
-		"@types/node": "^22.0.0",
-		"typescript": "^5.6.2",
-		"husky": "^9.0.11",
+                "@modelcontextprotocol/sdk": "^1.0.1",
+                "puppeteer": "^23.4.0",
+                "express": "^4.19.2"
+        },
+        "devDependencies": {
+                "@types/node": "^22.0.0",
+                "typescript": "^5.6.2",
+                "@types/express": "^4.17.21",
+                "husky": "^9.0.11",
+                "tsx": "^4.15.1",
 		"semantic-release": "^24.0.0",
 		"@semantic-release/changelog": "^6.0.3",
 		"@semantic-release/git": "^10.0.1",
